fix: extract get_host_port to Flask-free config.py and fix open() patching in tests
- Create src/html2md/config.py with get_host_port() and DEFAULT_PORT so tests
  can run without Flask being installed (removes pytest.importorskip skip guard)
- Update app.py to import from html2md.config instead of defining its own copy
- Update test_app.py to import from html2md.config (no Flask dependency)
- Fix test_cli_exceptions.py lines 40 and 62: add create=True to
  patch('html2md.cli.open') so the patch injects into the module namespace
  correctly instead of raising AttributeError

Agent-Logs-Url: https://github.com/badMadeC0/html2md/sessions/d986bc8f-f17a-49fe-8836-8cf713bc7b5a
